Swapping to react query
